+++ src/dotnetCampus.SourceYard/PackFlow/NuspecFiles/NuspecContexts/NuspecDependency.cs
