+++ src/dotnetCampus.SourceYard/PackFlow/NuspecFiles/NuspecContexts/NuspecFrameworkAssembly.cs
