+++ src/dotnetCampus.SourceYard/PackFlow/NuspecFiles/NuspecContexts/NuspecGroup.cs
