+++ src/dotnetCampus.SourceYard/PackFlow/NuspecFiles/NuspecContexts/NuspecMetadata.cs
