+++ src/dotnetCampus.SourceYard/PackFlow/NuspecFiles/NuspecContexts/NuspecPackage.cs
